[ci-helm] Deploy the DSS using Helm (#983)
* [ci] Update the CI to deploy the DSS using the helm chart

* Comments

+++ deploy/operations/ci/aws-1/test.sh
@@ -12,15 +12,50 @@ else
 fi
 cd "${BASEDIR}" || exit 1
 
-clean () {
-  echo "Cleaning infrastructure"
-  terraform destroy -auto-approve
-}
-
+# Initialize terraform
 terraform init
-clean
+# TODO: Fail if env is not clean
+
+# Deploy the Kubernetes cluster
 terraform apply -auto-approve
-# TODO: Deploy the DSS
+KUBE_CONTEXT="$(terraform output -raw cluster_context)"
+WORKSPACE_LOCATION="$(terraform output -raw workspace_location)"
+
+# Login into the Kubernetes Cluster
+cd "${WORKSPACE_LOCATION}"
+./get-credentials.sh
+aws sts get-caller-identity
+
+# Allow access to the cluster to AWS admins
+kubectl apply -f "aws_auth_config_map.yml"
+
+# Generate cockroachdb certificates
+./make-certs.sh
+./apply-certs.sh
+
+# Install the DSS using the helm chart
+cd "$BASEDIR/../../../services/helm-charts/dss"
+RELEASE_NAME="dss"
+helm dep update --kube-context="$KUBE_CONTEXT"
+helm upgrade --install --kube-context="$KUBE_CONTEXT" -f "${WORKSPACE_LOCATION}/helm_values.yml" "$RELEASE_NAME" .
+
 # TODO: Test the deployment of the DSS
-clean
+
+if [ -n "$DO_NOT_DESTROY" ]; then
+  echo "Destroy disabled. Exit."
+  exit 0
+fi
+
+# Cleanup
+# Delete workloads
+helm uninstall --kube-context="$KUBE_CONTEXT" "$RELEASE_NAME"
+
+# Delete PVC to delete persistant volumes
+kubectl delete pvc --all=true
+# TODO: Check completness
+
+# Delete cluster
+cd "$BASEDIR"
+terraform destroy -auto-approve
+
